Added an interface for mysql tables
Adds functionality to interface with mysql almost completely indirectly. Requires 2 functions for he entry, to_vec_string, which converts the data to a vector of strings, and from_mysql, which requires the entry to step through a vec of values and assign them to the struct. These functions have not been documented or pushed yet.